+++ GestorPlanta.java
@@ -8,7 +8,7 @@
 public class GestorPlanta extends Trabajador {
 
     public GestorPlanta(String nombre, String apellidos, String direccion, String DNI, 
-                        int numSeguridadSocial, int salario, String fechaIngreso) {
+                        int numSeguridadSocial, double salario, String fechaIngreso) {
         
         // Llamada al constructor de la clase abstracta Trabajador
         super(nombre, apellidos, direccion, DNI, numSeguridadSocial,"Gestor Planta", salario, fechaIngreso);
